// starport-ui/src/store/generated/cosmos/cosmos-sdk/ibc.core.connection.v1/module/rest.ts
export interface RpcStatus {
  /** @format int32 */
  code?: number;
  message?: string;
  details?: ProtobufAny[];
}
// ...
export interface V1QueryClientConnectionsResponse {
  /** slice of all the connection paths associated with a client. */
  connectionPaths?: string[];

  /** @format byte */
  proof?: string;

  /**
   * Normally the RevisionHeight is incremented at each height while keeping RevisionNumber
   * the same. However some consensus algorithms may choose to reset the
   * height in certain conditions e.g. hard forks, state-machine breaking changes
   * In these cases, the RevisionNumber is incremented so that height continues to
   * be monitonically increasing even as the RevisionHeight gets reset
   */
  proofHeight?: V1Height;
}

export interface V1QueryConnectionClientStateResponse {
  /**
   * IdentifiedClientState defines a client state with an additional client
   * identifier field.
   */
  identifiedClientState?: V1IdentifiedClientState;

  /** @format byte */
  proof?: string;

  /**
   * Normally the RevisionHeight is incremented at each height while keeping RevisionNumber
   * the same. However some consensus algorithms may choose to reset the
   * height in certain conditions e.g. hard forks, state-machine breaking changes
   * In these cases, the RevisionNumber is incremented so that height continues to
   * be monitonically increasing even as the RevisionHeight gets reset
   */
  proofHeight?: V1Height;
}
// ...
export interface V1QueryConnectionResponse {
  /**
   * ConnectionEnd defines a stateful object on a chain connected to another
   * separate one.
   * NOTE: there must only be 2 defined ConnectionEnds to establish
   * a connection between two chains.
   */
  connection?: V1ConnectionEnd;

  /** @format byte */
  proof?: string;

  /**
   * Normally the RevisionHeight is incremented at each height while keeping RevisionNumber
   * the same. However some consensus algorithms may choose to reset the
   * height in certain conditions e.g. hard forks, state-machine breaking changes
   * In these cases, the RevisionNumber is incremented so that height continues to
   * be monitonically increasing even as the RevisionHeight gets reset
   */
  proofHeight?: V1Height;
}

/**
* QueryConnectionsResponse is the response type for the Query/Connections RPC
method.
*/
export interface V1QueryConnectionsResponse {
  /** list of stored connections of the chain. */
  connections?: V1IdentifiedConnection[];

  /**
   * PageResponse is to be embedded in gRPC response messages where the
   * corresponding request message has used PageRequest.
   *
   *  message SomeResponse {
   *          repeated Bar results = 1;
   *          PageResponse page = 2;
   *  }
   */
  pagination?: V1Beta1PageResponse;

  /**
   * Normally the RevisionHeight is incremented at each height while keeping RevisionNumber
   * the same. However some consensus algorithms may choose to reset the
   * height in certain conditions e.g. hard forks, state-machine breaking changes
   * In these cases, the RevisionNumber is incremented so that height continues to
   * be monitonically increasing even as the RevisionHeight gets reset
   */
  height?: V1Height;
}
// ...
export type QueryParamsType = Record<string | number, any>;
export type ResponseFormat = keyof Omit<Body, "body" | "bodyUsed">;

export interface FullRequestParams extends Omit<RequestInit, "body"> {
  /** set parameter to `true` for call `securityWorker` for this request */
  secure?: boolean;
  /** request path */
  path: string;
  /** content type of request body */
  type?: ContentType;
  /** query params */
  query?: QueryParamsType;
  /** format of response (i.e. response.json() -> format: "json") */
  format?: keyof Omit<Body, "body" | "bodyUsed">;
  /** request body */
  body?: unknown;
  /** base url */
  baseUrl?: string;
  /** request cancellation token */
  cancelToken?: CancelToken;
}

export type RequestParams = Omit<FullRequestParams, "body" | "method" | "query" | "path">;

export interface ApiConfig<SecurityDataType = unknown> {
  baseUrl?: string;
  baseApiParams?: Omit<RequestParams, "baseUrl" | "cancelToken" | "signal">;
  securityWorker?: (securityData: SecurityDataType) => RequestParams | void;
}

export interface HttpResponse<D extends unknown, E extends unknown = unknown> extends Response {
  data: D;
  error: E;
}

type CancelToken = Symbol | string | number;

export enum ContentType {
  Json = "application/json",
  FormData = "multipart/form-data",
  UrlEncoded = "application/x-www-form-urlencoded",
}

export class HttpClient<SecurityDataType = unknown> {
  public baseUrl: string = "";
  private securityData: SecurityDataType = null as any;
  private securityWorker: null | ApiConfig<SecurityDataType>["securityWorker"] = null;
  private abortControllers = new Map<CancelToken, AbortController>();

  private baseApiParams: RequestParams = {
    credentials: "same-origin",
    headers: {},
    redirect: "follow",
    referrerPolicy: "no-referrer",
  };

  constructor(apiConfig: ApiConfig<SecurityDataType> = {}) {
    Object.assign(this, apiConfig);
  }

  public setSecurityData = (data: SecurityDataType) => {
    this.securityData = data;
  };

  private addQueryParam(query: QueryParamsType, key: string) {
    const value = query[key];

    return (
      encodeURIComponent(key) +
      "=" +
      encodeURIComponent(Array.isArray(value) ? value.join(",") : typeof value === "number" ? value : `${value}`)
    );
  }

  protected toQueryString(rawQuery?: QueryParamsType): string {
    const query = rawQuery || {};
    const keys = Object.keys(query).filter((key) => "undefined" !== typeof query[key]);
    return keys
      .map((key) =>
        typeof query[key] === "object" && !Array.isArray(query[key])
          ? this.toQueryString(query[key] as QueryParamsType)
          : this.addQueryParam(query, key),
      )
      .join("&");
  }

  protected addQueryParams(rawQuery?: QueryParamsType): string {
    const queryString = this.toQueryString(rawQuery);
    return queryString ? `?${queryString}` : "";
  }

  private contentFormatters: Record<ContentType, (input: any) => any> = {
    [ContentType.Json]: (input: any) =>
      input !== null && (typeof input === "object" || typeof input === "string") ? JSON.stringify(input) : input,
    [ContentType.FormData]: (input: any) =>
      Object.keys(input || {}).reduce((data, key) => {
        data.append(key, input[key]);
        return data;
      }, new FormData()),
    [ContentType.UrlEncoded]: (input: any) => this.toQueryString(input),
  };

  private mergeRequestParams(params1: RequestParams, params2?: RequestParams): RequestParams {
    return {
      ...this.baseApiParams,
      ...params1,
      ...(params2 || {}),
      headers: {
        ...(this.baseApiParams.headers || {}),
        ...(params1.headers || {}),
        ...((params2 && params2.headers) || {}),
      },
    };
  }

  private createAbortSignal = (cancelToken: CancelToken): AbortSignal | undefined => {
    if (this.abortControllers.has(cancelToken)) {
      const abortController = this.abortControllers.get(cancelToken);
      if (abortController) {
        return abortController.signal;
      }
      return void 0;
    }

    const abortController = new AbortController();
    this.abortControllers.set(cancelToken, abortController);
    return abortController.signal;
  };

  public abortRequest = (cancelToken: CancelToken) => {
    const abortController = this.abortControllers.get(cancelToken);

    if (abortController) {
      abortController.abort();
      this.abortControllers.delete(cancelToken);
    }
  };

  public request = <T = any, E = any>({
    body,
    secure,
    path,
    type,
    query,
    format = "json",
    baseUrl,
    cancelToken,
    ...params
  }: FullRequestParams): Promise<HttpResponse<T, E>> => {
    const secureParams = (secure && this.securityWorker && this.securityWorker(this.securityData)) || {};
    const requestParams = this.mergeRequestParams(params, secureParams);
    const queryString = query && this.toQueryString(query);
    const payloadFormatter = this.contentFormatters[type || ContentType.Json];

    return fetch(`${baseUrl || this.baseUrl || ""}${path}${queryString ? `?${queryString}` : ""}`, {
      ...requestParams,
      headers: {
        ...(type && type !== ContentType.FormData ? { "Content-Type": type } : {}),
        ...(requestParams.headers || {}),
      },
      signal: cancelToken ? this.createAbortSignal(cancelToken) : void 0,
      body: typeof body === "undefined" || body === null ? null : payloadFormatter(body),
    }).then(async (response) => {
      const r = response as HttpResponse<T, E>;
      r.data = (null as unknown) as T;
      r.error = (null as unknown) as E;

      const data = await response[format]()
        .then((data) => {
          if (r.ok) {
            r.data = data;
          } else {
            r.error = data;
          }
          return r;
        })
        .catch((e) => {
          r.error = e;
          return r;
        });

      if (cancelToken) {
        this.abortControllers.delete(cancelToken);
      }

      if (!response.ok) throw data;
      return data;
    });
  };
}

/**
 * @title ibc/core/connection/v1/tx.proto
 * @version version not set
 */
export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDataType> {
  /**
 * No description
 * 
 * @tags Query
 * @name QueryClientConnections
 * @summary ClientConnections queries the connection paths associated with a client
state.
 * @request GET:/ibc/core/connection/v1beta1/client_connections/{clientId}
 */
  queryClientConnections = (clientId: string, params: RequestParams = {}) =>
    this.request<V1QueryClientConnectionsResponse, RpcStatus>({
      path: `/ibc/core/connection/v1beta1/client_connections/${clientId}`,
      method: "GET",
      format: "json",
      ...params,
    });

  /**
   * No description
   *
   * @tags Query
   * @name QueryConnections
   * @summary Connections queries all the IBC connections of a chain.
   * @request GET:/ibc/core/connection/v1beta1/connections
   */
  queryConnections = (
    query?: {
      "pagination.key"?: string;
      "pagination.offset"?: string;
      "pagination.limit"?: string;
      "pagination.countTotal"?: boolean;
    },
    params: RequestParams = {},
  ) =>
    this.request<V1QueryConnectionsResponse, RpcStatus>({
      path: `/ibc/core/connection/v1beta1/connections`,
      method: "GET",
      query: query,
      format: "json",
      ...params,
    });

  /**
   * No description
   *
   * @tags Query
   * @name QueryConnection
   * @summary Connection queries an IBC connection end.
   * @request GET:/ibc/core/connection/v1beta1/connections/{connectionId}
   */
  queryConnection = (connectionId: string, params: RequestParams = {}) =>
    this.request<V1QueryConnectionResponse, RpcStatus>({
      path: `/ibc/core/connection/v1beta1/connections/${connectionId}`,
      method: "GET",
      format: "json",
      ...params,
    });

  /**
 * No description
 * 
 * @tags Query
 * @name QueryConnectionClientState
 * @summary ConnectionClientState queries the client state associated with the
connection.
 * @request GET:/ibc/core/connection/v1beta1/connections/{connectionId}/client_state
 */
  queryConnectionClientState = (connectionId: string, params: RequestParams = {}) =>
    this.request<V1QueryConnectionClientStateResponse, RpcStatus>({
      path: `/ibc/core/connection/v1beta1/connections/${connectionId}/client_state`,
      method: "GET",
      format: "json",
      ...params,
    });

  /**
 * No description
 * 
 * @tags Query
 * @name QueryConnectionConsensusState
 * @summary ConnectionConsensusState queries the consensus state associated with the
connection.
 * @request GET:/ibc/core/connection/v1beta1/connections/{connectionId}/consensus_state/revision/{revisionNumber}/height/{revisionHeight}
 */
  queryConnectionConsensusState = (
    connectionId: string,
    revisionNumber: string,
    revisionHeight: string,
    params: RequestParams = {},
  ) =>
    this.request<V1QueryConnectionConsensusStateResponse, RpcStatus>({
      path: `/ibc/core/connection/v1beta1/connections/${connectionId}/consensus_state/revision/${revisionNumber}/height/${revisionHeight}`,
      method: "GET",
      format: "json",
      ...params,
    });
}